+++ components/space/Space.razor
@@ -5,7 +5,7 @@
     SpaceItemCount = 0;
 }
 <CascadingValue Value="this" IsFixed="@true">
-    <div class="@ClassMapper.Class" style="@Style @InnerStyle" id="@Id" @ref="@Ref">
+    <div class="@ClassMapper.Class" style="@Style @InnerStyle @_gapStyle" id="@Id" @ref="@Ref">
         @ChildContent
     </div>
 </CascadingValue>
