+++ components/space/SpaceItem.razor
@@ -1,7 +1,6 @@
 ﻿@namespace AntDesign
 @inherits AntDomComponentBase
 
-@{ ChangeSize(); }
 @if (Parent?.Split != null && _index != 0)
 {
     <span class="ant-space-item-split" style="margin-right: 4px;">
@@ -9,6 +8,6 @@
     </span>
 }
 
-<div class="@ClassMapper.Class" style="@Style @_marginStyle" @ref="Ref" id="@Id">
+<div class="@ClassMapper.Class" style="@Style" @ref="Ref" id="@Id">
     @ChildContent
 </div>
