+++ components/space/SpaceItem.razor.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections.Generic;
 using Microsoft.AspNetCore.Components;
 
 namespace AntDesign
@@ -21,14 +20,6 @@ namespace AntDesign
         [Parameter]
         public RenderFragment ChildContent { get; set; }
 
-        private static readonly Dictionary<SpaceSize, string> _spaceSize = new()
-        {
-            [SpaceSize.Small] = "8",
-            [SpaceSize.Middle] = "16",
-            [SpaceSize.Large] = "24"
-        };
-
-        private string _marginStyle = "";
         private int _index;
 
         protected override void OnInitialized()
@@ -42,42 +33,6 @@ namespace AntDesign
         {
             _index = Parent.SpaceItemCount++;
             base.OnParametersSet();
-        }
-
-        private void ChangeSize()
-        {
-            var size = Parent.Size;
-            var direction = Parent.Direction;
-
-            size.Switch(singleSize =>
-            {
-                _marginStyle = direction == SpaceDirection.Horizontal ? (_index != Parent.SpaceItemCount - 1 ? $"margin-right:{GetSize(singleSize)};" : "") : $"margin-bottom:{GetSize(singleSize)};";
-            },
-            singleSize =>
-            {
-                _marginStyle = direction == SpaceDirection.Horizontal ? (_index != Parent.SpaceItemCount - 1 ? $"margin-right:{GetSize(singleSize)};" : "") : $"margin-bottom:{GetSize(singleSize)};";
-            },
-            arraySize =>
-            {
-                _marginStyle = (_index != Parent.SpaceItemCount - 1 ? $"margin-right:{GetSize(arraySize.Item1)};" : "") + $"margin-bottom:{GetSize(arraySize.Item2)};";
-            });
-        }
-
-        private CssSizeLength GetSize(SpaceSize size)
-        {
-            var originalSize = _spaceSize[size];
-
-            return GetSize(originalSize);
-        }
-
-        private CssSizeLength GetSize(string size)
-        {
-            if (Parent?.Split != null)
-            {
-                return ((CssSizeLength)size).Value / 2;
-            }
-
-            return size;
         }
     }
 }
